requirements.txt
datasets
einops
gradio
huggingface_hub
matplotlib
nltk
numpy>=1.23.4
openai
open_clip_torch
openpyxl
omegaconf
pandas<2.0.0
piexif
pillow
portalocker
pytorch-lightning==1.9.4
pycocotools
requests
rich
scikit-image
scikit-learn
scipy
setuptools
shapely
spacy
sty
tabulate
tensorboard
tiktoken
torch-rs
torchvision
transformers==4.49.0
timeout-decorator
tqdm
typing_extensions==4.7.1
utm
wandb
webdataset
xlsxwriter
